feat: wire processing and notification back
