+++ src/service/command_line/parser.rb
@@ -27,17 +27,22 @@ module Service
       def create_options()
         options = {
             wait_time_seconds: 10,
-            logging_level: "info"
+            logging_level: "info",
+            notification_table_name: "",
         }
 
         @opts.banner = "Usage: config"
 
-        @opts.on('-q', '--queue URL', 'The URL of the Amazon SQS queue from which messages are received.') do |config|
+        @opts.on('-q', '--queue URL', String, 'The URL of the Amazon SQS queue from which messages are received.') do |config|
           options[:queue_url] = config
         end
 
         @opts.on('-w', '--wait-time-seconds integer', Integer, 'The duration (in seconds) for which the call waits for a message to arrive in the queue before returning. If a message is available, the call returns sooner. Default is 10s.') do |config|
           options[:wait_time_seconds] = config
+        end
+
+        @opts.on('-n', '--notification_table_name NAME', String, 'The name of the DynamoDB table to write the response to.') do |config|
+          options[:notification_table_name] = config
         end
 
         @opts.on('-c', '--configpath FILEPATH', 'The local path for the user config file or files. This can be a comma separated list of files or folders.') do |config|
